+++ tests/integration/plugins/test_gradle.py
@@ -24,25 +24,24 @@ import yaml
 from craft_parts import LifecycleManager, Step
 from craft_parts.infos import ProjectInfo
 
-pytestmark = [pytest.mark.java]
+pytestmark = [pytest.mark.slow, pytest.mark.java]
 
 
 @pytest.fixture
-def testing_source_dir(new_dir):
-    source_location = Path(__file__).parent / "test_gradle"
-    shutil.copytree(source_location, new_dir, dirs_exist_ok=True)
-    return Path(new_dir)
-
-
-@pytest.fixture
-def use_gradlew(request, testing_source_dir):
+def use_gradlew(request, new_dir):
     if not (use_wrapper := request.param):
-        (testing_source_dir / "gradlew").unlink(missing_ok=True)
+        Path(new_dir, "gradlew").unlink(missing_ok=True)
     return use_wrapper
 
 
+def copy_tree(new_dir, copy_dir):
+    source_location = Path(__file__).parent / "test_gradle" / copy_dir
+    shutil.copytree(source_location, new_dir, dirs_exist_ok=True)
+
+
 @pytest.mark.parametrize("use_gradlew", [True, False], indirect=True)
-def test_gradle_plugin(new_dir, testing_source_dir, partitions, use_gradlew):
+def test_gradle_plugin(new_dir, partitions, use_gradlew):
+    copy_tree(new_dir, "simple")
     part_name = "foo"
     parts_yaml = textwrap.dedent(
         f"""
@@ -51,7 +50,7 @@ def test_gradle_plugin(new_dir, testing_source_dir, partitions, use_gradlew):
             plugin: gradle
             gradle-task: testWrite build
             gradle-init-script: init.gradle
-            source: {testing_source_dir}
+            source: {new_dir}
             build-packages: [openjdk-21-jdk]
             build-snaps:
                 - gradle
@@ -94,3 +93,39 @@ def _test_core_gradle_plugin_build_output(project_info: ProjectInfo) -> None:
         [str(java_binary), "-jar", f"{prime_dir}/jar/build-1.0.jar"], text=True
     )
     assert output.strip() == "Hello from Gradle-built Java"
+
+
+def test_gradle_self_contained(new_dir, partitions):
+    copy_tree(new_dir, "self-contained")
+    parts = yaml.safe_load((new_dir / "parts.yaml").read_text(encoding="utf-8"))
+
+    lf = LifecycleManager(
+        parts,
+        application_name="test_gradle_self_contained",
+        cache_dir=new_dir,
+        work_dir=new_dir,
+        partitions=partitions,
+    )
+    actions = lf.plan(Step.PRIME)
+    with lf.action_executor() as ctx:
+        ctx.execute(actions)
+
+    backstage = lf.project_info.backstage_dir / "maven-use"
+    assert backstage.is_dir()
+    # assert dependencies were published to backstage
+    for part in ["hello", "hello-plugin"]:
+        assert list(backstage.rglob(f"{part}-1.0.0.jar"))
+        assert list(backstage.rglob(f"{part}-1.0.0.pom"))
+
+    assert list(backstage.rglob("org.starcraft.hello-plugin.gradle.plugin"))
+
+    output = subprocess.check_output(
+        [
+            "java",
+            "-cp",
+            f"{lf.project_info.prime_dir}/jar/*",
+            "org.starcraft.HelloCraft",
+        ],
+        text=True,
+    )
+    assert output.strip() == "Hello, craft!"