+++ docs/reference/changelog.rst
@@ -27,6 +27,7 @@ Changelog
 
 New features:
 
+- Add self-contained support for parts using the Gradle plugin.
 - The Gradle daemon is now disabled by default when using the Gradle plugin. You can
   control this behaviour with the new ``gradle-use-daemon`` key.